backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
rembg>=2.0.59
Pillow>=10.4.0
onnxruntime>=1.23.0
